::Hash now generates a uint64_t instead of a int32_t

+++ src/common/types/hash.cpp
@@ -6,21 +6,21 @@ using namespace std;
 
 namespace duckdb {
 
-template <> int32_t Hash(uint64_t val) {
+template <> uint64_t Hash(uint64_t val) {
 	return murmurhash64((uint32_t *)&val);
 }
 
-template <> int32_t Hash(int64_t val) {
+template <> uint64_t Hash(int64_t val) {
 	return murmurhash64((uint32_t *)&val);
 }
 
-template <> int32_t Hash(double val) {
+template <> uint64_t Hash(double val) {
 	return murmurhash64((uint32_t *)&val);
 }
 
-template <> int32_t Hash(const char *str) {
-	uint32_t hash = 5381;
-	uint32_t c;
+template <> uint64_t Hash(const char *str) {
+	uint64_t hash = 5381;
+	uint64_t c;
 
 	while ((c = *str++)) {
 		hash = ((hash << 5) + hash) + c;
@@ -29,7 +29,7 @@ template <> int32_t Hash(const char *str) {
 	return hash;
 }
 
-template <> int32_t Hash(char *val) {
+template <> uint64_t Hash(char *val) {
 	return Hash<const char *>(val);
 }
